Исправлено индивидуальное задание

+++ РyCharm/individual.py
@@ -2,9 +2,9 @@ import math
 
 print('Enter the 3 sides of the triangle')
 
-z1 = int(input())
-z2 = int(input())
-z3 = int(input())
+z1 = float(input())
+z2 = float(input())
+z3 = float(input())
 
 p = (z1+z2+z3) / 2
 st = math.sqrt(p*(p-z1)*(p-z2)*(p-z3))
